requirements.txt
ephem>=4.1.0
numpy>=1.21.0
reportlab>=4.0.0
Pillow>=9.0.0
requests>=2.28.0
